libocfs2: Remove error code OCFS2_ET_DIRENT_NOT_FOUND
commit 3de2f0e3e643b6ee03c481f8e6ec3eba30f2a94e added error code
OCFS2_ET_DIRENT_NOT_FOUND which appears the same as the existing error code
OCFS2_ET_FILE_NOT_FOUND. The former was added as part of the support for indexed
directories in tools.

Having two error codes for the saem error is problematic because ocfs2_lookup()
could return either of the two depending on whether indexing is enabled or not.
While we could teach all the users to expect both errors, I see no advantage
in that.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>