fsck.ocfs2: Read wall clock before and after reading resource usage
During computing resource usage, fsck's reading of the wall clock should
encompass its reading of the resource usage. This will ensure the wall time
is not less than the user time.

While this fix is good, I am still seeing wall time less than user time.
It is not a pressing bug as it only shows up when that values are close to zero.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>