Fix a memory leak problem in image.c
Whether OCFS2_FLAG_IMAGE_FILE is specified in 'fs_flags' or not, the
handler will allocate memory for 'ost_bmparr'.

In main(), there are two paths of memory allocation for 'ost_bmparr':
1. OCFS2_FLAG_IMAGE_FILE is specified: ocfs2_image_load_bitmap()->
   ocfs2_image_alloc_bitmap()->ocfs2_malloc0().
2. OCFS2_FLAG_IMAGE_FILE is not specified: main()->o2image_initialize()->
   ocfs2_image_alloc_bitmap()->ocfs2_malloc0().

But in ocfs2_image_free_bitmap(), 'ost_bmparr' would not be freed if
OCFS2_FLAG_IMAGE_FILE is not specified. So we should free 'ost_bmparr'
anyway.

Signed-off-by: Jun Piao <piaojun@huawei.com>
Reviewed-by: Alex Chen <alex.chen@huawei.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.de>