Add check to compare journal size and cluster size
If user doesn't provide right Journal size, mkfs should fail. Currently it
succeeds but ends up as unusable fs.

1. mkfs.ocfs2 -J size=4M -C 1M -N 2 -L xvdh --cluster-name=ocfs2big
  --cluster-stack=o2cb --force --global-heartbeat /dev/xvdh
2. mount -t ocfs2 /dev/xvdh /mnt

3. mkdir aa
    mkdir: cannot create directory `aa': No space left on device

This happens because of the following:

The kernel allows the maximum transaction buffer to be 1\4 th of the journal
size and this is further divided by 2 for transaction reservation support.
Some operations such as mkdir can require more than a cluster worth of journal
credits. Such operations will fail if the journal size is not greater than
cluster size * 8.

This patch adds this check for user provided values and modifies the default
calculation to account for the same. It will fail mkfs if the conditions are
not met instead of creating a non-functional filesystem.

Signed-off-by: Ashish Samant <ashish.samant@oracle.com>
Reviewed-by: Srinivas Eeda <srinivas.eeda@oracle.com>
Signed-off-by: Junxiao Bi <junxiao.bi@oracle.com>