fsck.ocfs2: fix journal tail wrapper
If there is one transaction, its last block but one located at the
tail of the journal and its last commit block located at the head
of the journal, Crash the kernel and when fsck.ocfs2 -a run during
the reboot, the journal replay will fail due to the wrapper error.
This will make all the following transaction starting from the journal
head lost, and this will corrupt the fs.

Acked-by: Srinivas Eeda <srinivas.eeda@oracle.com>
Signed-off-by: Junxiao Bi <junxiao.bi@oracle.com>