fsck.ocfs2: fix corruption when truncate reflink file
The following steps will cause a corruption:
1. reflink test test.ref
2. open test.ref
3. remove test.ref
4. crash the system
5. fsck.ocfs2 -a $ocfs2_volume

fsck to check, you will find the following error:

Pass 1: Checking inodes and blocks
[REFCOUNT_COUNT] Refcount tree at 6901 claims to have 2 files associated with it, but we only found 1.Update the count number? y
[REFCOUNT_COUNT_INVALID] clusters 9612289 with len 768 have 2 refcount while there are 1 files point to them. Correct the refcount value? y
[CLUSTER_ALLOC_BIT] Cluster 9612289 is in use but isn't set in the global cluster bitmap. Set its bit in the bitmap? y

The root cause of this issue is that refcount tree is not updated right when
truncating files in orphan dir. The refcount flag in extents have been memset
to zero when try to decrease refcount, so share cluster freed but not update
refcount tree. This causes corruption [REFCOUNT_COUNT_INVALID] and [CLUSTER_ALLOC_BIT].

[REFCOUNT_COUNT] error is because refcount tree's rf_count is decreased by 1
when file size is truncated to zero.

Signed-off-by: Junxiao Bi <junxiao.bi@oracle.com>
Acked-by: Ryan Ding<ryan.ding@oracle.com>