fsck.ocfs2: do not remove the chain when verifying global bitmap chain alloc
In verify_chain_alloc(), when cr->c_blkno is equal to 0 we think this
chain record is missing, we remove the chain and move the last chain into
this missing chain record. But it's not necessary to remove the missing
chain record when we verify the global bitmap chain allocator, otherwise,
the last chain record will be overwritten when we recover the missing
chain record in verify_bitmap_descs(), in this way, the last chain record
will be lost forever.

Signed-off-by: Alex Chen <alex.chen@huawei.com>
Signed-off-by: Joseph Qi <joseph.qi@huawei.com>