o2cb: unable to start o2cb service on 2.6.18-xxx kernel.
commit 4483004b5723("Auto setup o2cb stack as default if no stack is setup")
refactored the code and this bug was introduce. It does not realize that on
2.6.18-xxx kernel, there is no /sys/fs/ocfs2/cluster-stack synthetic file.
Fix is to realize that after loading modules, if the above file does not exist,
then use the classic stack as before. Since ocfs2_stack_o2cb don't exist
on 2.6.18-xxx, use the modprob quite option to to suppress printing
"FATAL: Module ocfs2_stack_o2cb not found."

Signed-off-by: Tariq Saeed <tariq.x.saeed@oracle.com>
Reviewed-by: Junxiao Bi <junxiao.bi@oracle.com>