tunefs.ocfs2: read journal superblock even for online operations
Currently, we do not read the journal superblock for tunefs operations
that can be done online. This leads to a problem when we are dynamically
increasing the size of the fs above 16TB. This is because the block64
incompat flag, which identifies whether the fs can support this feature,
is only stored on disk in the journal super block. Since we don't read the
journal super block for operations that can be done online, we can end up
wrongly assuming that the block64 feature is not enabled and fail the
resize operation.

Fix this by reading the journal superblock for online as well as offline
operations. Check dirty journals and bitmap only for offline operations.

Signed-off-by: Ashish Samant <ashish.samant@oracle.com>
Acked-by: Joseph Qi <jiangqi903@gmail.com>