fsck.ocfs2: do not try locking after replaying journals if -F is given
With -F option, `fsck.ocfs2` skips locking down the cluster, but not if
the journals are replayed. If it has to replay the journals then it
tries to lock down the cluster regardless of -F.

This fix moves the check to `fsck_lock_fs()` function so the flag
is checked in any case when cluster locking is attempted.

Signed-off-by: Konstantin Neshin <b.m.kast@gmail.com>